feat(operations): section 0 CREATE on Content tab — extracted ProjectCreateForm + RoutineCreateForm, entity-default wiring, onCreated refresh
